pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "legalos"
version = "0.1.0"
description = "AI-powered legal document analyzer for Indian startup founders"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.42.0",
    "click>=8.1",
    "pydantic>=2.0",
    "rich>=13.0",
    "pymupdf>=1.25.0",
    "pymupdf4llm>=0.0.17",
    "python-docx>=1.1.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
ocr = ["easyocr>=1.7"]

[project.scripts]
legalos = "legalos.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/legalos"]
